server/generate_painting/requirements.txt
Flask>=2.0
Flask-CORS>=3.0
Pillow>=9.0
google-generativeai>=0.5.0
gunicorn>=20.0 